requirements.txt
graphviz~=0.14
hypothesis~=5.20.3
psycopg2~=2.8.5
python-benedict~=0.18.1
pytest~=5.4.3
pytest-mock~=3.2.0
PyYAML~=5.3.1
